requirements.txt
streamlit
requests
python-whois
fpdf2
qrcode[pil]
mercadopago
